Move implementation to source file.
